+++ plugins/voxelman/clientdb/clientinfo.d
@@ -3,9 +3,8 @@ Copyright: Copyright (c) 2015-2016 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.server.clientinfo;
+module voxelman.clientdb.clientinfo;
 
-import netlib.connection : ENetPeer;
 import dlib.math.vector : vec3, vec2;
 
 import voxelman.core.config;