+++ plugins/voxelman/clientdb/plugin.d
@@ -1,0 +1,163 @@
+/**
+Copyright: Copyright (c) 2016 Andrey Penechko.
+License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
+Authors: Andrey Penechko.
+*/
+module voxelman.clientdb.plugin;
+
+import std.experimental.logger;
+import netlib;
+import pluginlib;
+
+import voxelman.core.config;
+import voxelman.core.events;
+import voxelman.net.events;
+import voxelman.core.packets;
+import voxelman.net.packets;
+import voxelman.storage.coordinates;
+
+import voxelman.eventdispatcher.plugin;
+import voxelman.net.plugin;
+import voxelman.world.plugin;
+
+import voxelman.clientdb.clientinfo;
+
+shared static this()
+{
+	pluginRegistry.regServerPlugin(new ClientDb);
+}
+
+final class ClientDb : IPlugin
+{
+private:
+	EventDispatcherPlugin evDispatcher;
+	NetServerPlugin connection;
+	ServerWorld serverWorld;
+
+public:
+	ClientInfo*[ClientId] clients;
+
+	// IPlugin stuff
+	mixin IdAndSemverFrom!(voxelman.clientdb.plugininfo);
+
+	override void registerResources(IResourceManagerRegistry resmanRegistry) {}
+	override void preInit() {}
+	override void init(IPluginManager pluginman)
+	{
+		evDispatcher = pluginman.getPlugin!EventDispatcherPlugin;
+		connection = pluginman.getPlugin!NetServerPlugin;
+		serverWorld = pluginman.getPlugin!ServerWorld;
+
+		evDispatcher.subscribeToEvent(&handleClientConnected);
+		evDispatcher.subscribeToEvent(&handleClientDisconnected);
+
+		connection.registerPacketHandler!LoginPacket(&handleLoginPacket);
+		connection.registerPacketHandler!ViewRadiusPacket(&handleViewRadius);
+		connection.registerPacketHandler!ClientPositionPacket(&handleClientPosition);
+	}
+
+	bool isLoggedIn(ClientId clientId)
+	{
+		ClientInfo* clientInfo = clients[clientId];
+		return clientInfo.isLoggedIn;
+	}
+
+	string[ClientId] clientNames()
+	{
+		string[ClientId] names;
+		foreach(id, client; clients) {
+			names[id] = client.name;
+		}
+
+		return names;
+	}
+
+	string clientName(ClientId clientId)
+	{
+		import std.string : format;
+		auto cl = clients.get(clientId, null);
+		return cl ? cl.name : format("%s", clientId);
+	}
+
+	auto loggedInClients()
+	{
+		import std.algorithm : filter, map;
+		return clients.byKeyValue.filter!(a=>a.value.isLoggedIn).map!(a=>a.value.id);
+	}
+
+	void spawnClient(vec3 pos, vec2 heading, ClientId clientId)
+	{
+		ClientInfo* info = clients[clientId];
+		info.pos = pos;
+		info.heading = heading;
+		connection.sendTo(clientId, ClientPositionPacket(pos, heading));
+		connection.sendTo(clientId, SpawnPacket());
+		updateObserverVolume(info);
+	}
+
+	void handleClientConnected(ref ClientConnectedEvent event)
+	{
+		clients[event.clientId] = new ClientInfo(event.clientId);
+		connection.sendTo(event.clientId, PacketMapPacket(connection.packetNames));
+	}
+
+	void handleClientDisconnected(ref ClientDisconnectedEvent event)
+	{
+		serverWorld.chunkObserverManager.removeObserver(event.clientId);
+
+		infof("%s %s disconnected", event.clientId,
+			clients[event.clientId].name);
+
+		connection.sendToAll(ClientLoggedOutPacket(event.clientId));
+		clients.remove(event.clientId);
+	}
+
+	void updateObserverVolume(ClientInfo* info)
+	{
+		if (info.isLoggedIn) {
+			ChunkWorldPos chunkPos = BlockWorldPos(info.pos);
+			serverWorld.chunkObserverManager.changeObserverVolume(info.id, chunkPos, info.viewRadius);
+		}
+	}
+
+	void handleLoginPacket(ubyte[] packetData, ClientId clientId)
+	{
+		LoginPacket packet = unpackPacket!LoginPacket(packetData);
+		ClientInfo* info = clients[clientId];
+		info.name = packet.clientName;
+		info.id = clientId;
+		info.isLoggedIn = true;
+		spawnClient(info.pos, info.heading, clientId);
+
+		infof("%s %s logged in", clientId, clients[clientId].name);
+
+		connection.sendTo(clientId, SessionInfoPacket(clientId, clientNames));
+		connection.sendToAllExcept(clientId, ClientLoggedInPacket(clientId, packet.clientName));
+
+		evDispatcher.postEvent(ClientLoggedInEvent(clientId));
+	}
+
+	void handleViewRadius(ubyte[] packetData, ClientId clientId)
+	{
+		import std.algorithm : clamp;
+		auto packet = unpackPacket!ViewRadiusPacket(packetData);
+		infof("Received ViewRadiusPacket(%s)", packet.viewRadius);
+		ClientInfo* info = clients[clientId];
+		info.viewRadius = clamp(packet.viewRadius,
+			MIN_VIEW_RADIUS, MAX_VIEW_RADIUS);
+		updateObserverVolume(info);
+	}
+
+	void handleClientPosition(ubyte[] packetData, ClientId clientId)
+	{
+		if (isLoggedIn(clientId))
+		{
+			auto packet = unpackPacket!ClientPositionPacket(packetData);
+			ClientInfo* info = clients[clientId];
+			info.pos = packet.pos;
+			info.heading = packet.heading;
+			updateObserverVolume(info);
+		}
+	}
+
+}