+++ plugins/voxelman/clientdb/plugininfo.d
@@ -1,0 +1,6 @@
+module voxelman.clientdb.plugininfo;
+enum id = "voxelman.clientdb";
+enum semver = "0.6.0";
+enum deps = [];
+enum clientdeps = [];
+enum serverdeps = [];